Short-circuit fitting and summary with learn()
Upon testing, it was unneccessary to separate the tasks of (1) using learn() to fit all models and save them to a `fit` element in pipelearner object; and to then (2) extract a summary table of these fits with summary()/recover_fits(). These two steps seemed to always be occuring together. To clean up  this workflow, these steps are now combined in learn(). I.e., setup pipelearner object and, when ready, call learn() to return a tibble of fitted models and all other relevant info.